+++ Sources/MapboxNavigation/SpeedLimitView.swift
@@ -2,14 +2,6 @@ import UIKit
 import MapboxDirections
 
 /**
- Indicates which view mode to display, the current speed or the speed limit.
- */
-enum SpeedViewMode {
-    case speed
-    case speedLimit
-}
-
-/**
  A view that displays a speed limit and resembles a real-world speed limit sign.
  */
 public class SpeedLimitView: UIView {
@@ -47,18 +39,6 @@ public class SpeedLimitView: UIView {
     // MARK: Populating the Sign
     
     /**
-     The speed view mode to display. Defaults to `.speedLimit`.
-     */
-    var speedViewMode: SpeedViewMode = .speedLimit {
-        // TODO: After added the function to draw speed view, make this property public.
-        didSet {
-            if speedViewMode != oldValue {
-                update()
-            }
-        }
-    }
-    
-    /**
      The speed limit to display.
      
      The view displays the value of this property as is without converting it to another unit.
@@ -76,13 +56,7 @@ public class SpeedLimitView: UIView {
      
      The view displays the value of this property as is without converting it to another unit.
      */
-    public var speed: Measurement<UnitSpeed>? {
-        didSet {
-            if speed != oldValue {
-                update()
-            }
-        }
-    }
+    public var currentSpeed: Measurement<UnitSpeed>?
     
     /**
      The sign standard that specifies the design that the view depicts.
@@ -149,19 +123,7 @@ public class SpeedLimitView: UIView {
     override public func draw(_ rect: CGRect) {
         super.draw(rect)
         
-        switch speedViewMode {
-        case .speedLimit:
-            drawSpeedLimit()
-        case .speed:
-            // TODO: Add the function to draw speed view.
-            break
-        }
-    }
-    
-    func drawSpeedLimit() {
-        guard let signStandard = signStandard else {
-            return
-        }
+        guard let signStandard = signStandard else { return }
         
         let formattedSpeedLimit: String